Ignore *.png and *.jpg in vim
In order to not make Command-T's head explode with thousands of image
files, let's go ahead and ignore them in vim. The problem I was seeing
was CT wasn't showing my spec or vendor folder because the application
was generating fake data in public/uploads and causing thousands or
images. That was OK, and expected (although in production they'll end
up on S3 or another CDN).

The problem was CT was including these files
and they were filling up the g:CommandTMaxFiles count with a bunch of
files that wouldn't get edited in Vim. Ignoring them lowered the number
of files CT indexed and then spec and vendor started showing up again.